refactor: pass readability ratchet and clippy gates
